tubaro-comments: Sort the messages by pub_date
The messages in the RSS feed appear in reverse order. Previously that
would mean that if multiple messages are found between invocations of
the script then they would also appear in reverse order in the feed.
Instead of sending them directly they are now first collected into an
array with the pub_date and sorted before sending.